tp3/requirements.txt
python-dotenv
langchain
langgraph
langchain-core
langchain-openai
langchain-pinecone
langchain-text-splitters
pinecone
dash
dash-bootstrap-components
